Renamed test class
